+++ test/test_collada.py
@@ -267,7 +267,142 @@ class TestCOLLADA(EnvironmentSetup):
         assert(env2.Load('test_externalref_joints.dae'))
         misc.CompareBodies(robot,env2.GetRobots()[0])
         assert(len(env.GetBodies())==len(env2.GetBodies()))
-        
+
+    def test_externalref_scene(self):
+        xmldata = '''<Environment>
+  <Robot file="robots/barrett-wamhand.dae" name="BarrettWAM">
+    <translation>-0.8 0.14 1</translation>
+  </Robot>
+
+  <KinBody name="floorwalls">
+    <Body name="basefloor" type="static">
+      <Translation>0 0 0</Translation>
+      <Geom type="box">
+        <extents>2.5 2.5 0.005</extents>
+        <translation>0 0 -0.005</translation>
+        <diffuseColor>.6 .6 .6</diffuseColor>
+      	<ambientColor>0.6 0.6 0.6</ambientColor>
+      </Geom>
+      <Geom type="box">
+        <extents>2.5 0.01 0.2</extents>
+        <translation>0 -2.5 0.2</translation>
+        <diffuseColor>.6 .6 .6</diffuseColor>
+      	<ambientColor>0.6 0.6 0.6</ambientColor>
+      </Geom>
+      <Geom type="box">
+        <extents>2.5 0.01 0.2</extents>
+        <translation>0 2.5 0.2</translation>
+        <diffuseColor>.6 .6 .6</diffuseColor>
+      	<ambientColor>0.6 0.6 0.6</ambientColor>
+      </Geom>
+      <Geom type="box">
+        <extents>0.01 2.5 0.2</extents>
+        <translation>2.5 0 0.2</translation>
+        <diffuseColor>.6 .6 .6</diffuseColor>
+      	<ambientColor>0.6 0.6 0.6</ambientColor>
+      </Geom>
+      <Geom type="box">
+        <extents>0.01 2.5 0.2</extents>
+        <translation>-2.5 0 0.2</translation>
+        <diffuseColor>.6 .6 .6</diffuseColor>
+      	<ambientColor>0.6 0.6 0.6</ambientColor>
+      </Geom>
+    </Body>
+  </KinBody>
+  <KinBody name="pole">
+    <translation>-0.312 0.416 1</translation>
+    <Body name="basepole" type="static">
+      <Geom type="box">
+        <extents>0.05 0.05 1</extents>
+        <diffuseColor>1 .2 .2</diffuseColor>
+      </Geom>
+    </Body>
+  </KinBody>
+  <KinBody name="pole2">
+    <translation>1.3 0.6 1</translation>
+    <Body name="basepole" type="static">
+      <Geom type="box">
+        <extents>0.05 0.05 1</extents>
+        <diffuseColor>1 .2 .2</diffuseColor>
+      </Geom>
+    </Body>
+  </KinBody>
+  <KinBody name="pole3">
+    <translation>0.8 -0.9 1</translation>
+    <Body name="basepole" type="static">
+      <Geom type="box">
+        <extents>0.05 0.05 1</extents>
+        <diffuseColor>1 .2 .2</diffuseColor>
+      </Geom>
+    </Body>
+  </KinBody>
+
+  <KinBody name="wall1">
+    <Translation>-0.173 -0.247 0</Translation>
+    <Body name="basewall" type="static">
+      <Geom type="box">
+        <Translation>-0.735 -0.80 0.675</Translation>
+        <extents>0.07 0.71 0.73</extents>
+        <diffuseColor>.64 .64 .64</diffuseColor>
+      </Geom>
+      <Geom type="box">
+        <Translation>-1.235 -0.46 0.7</Translation>
+        <extents>0.54 0.35 0.06</extents>
+        <diffuseColor>.64 .64 .64</diffuseColor>
+      </Geom>
+      <Geom type="box">
+        <Translation>-1.235 -0.76 0.675</Translation>
+        <extents>0.54 0.06 0.71</extents>
+        <diffuseColor>.64 .64 .64</diffuseColor>
+      </Geom>
+    </Body>
+  </KinBody>
+  
+  <KinBody name="mug1" file="data/mug1.dae">
+    <Translation> -0.0568 -0.2406 0.7550</Translation>
+  </KinBody>
+  <KinBody name="mug2" file="data/mug1.dae">
+    <Translation>-0.02   0.15    0.7550</Translation>
+  </KinBody>
+  <KinBody name="mug3" file="data/mug1.dae">
+    <Translation>0.0854  -0.0769    0.7550</Translation>
+  </KinBody>
+  <KinBody name="mug4" file="data/mug1.dae">
+    <Translation>-0.23   -0.1446    0.7550</Translation>
+  </KinBody>
+  <KinBody name="mug5" file="data/mug1.dae">
+    <Translation>-0.1   -0.0151    0.7550</Translation>
+  </KinBody>
+  <KinBody name="mug6" file="data/mug1.dae">
+    <Translation>-0.2   0.11    0.7550</Translation>
+  </KinBody>
+
+  <KinBody name="table">
+    <Translation>-0.4966 1.0164 0.9</Translation>
+    <body name="basetable">
+      <geom type="box">
+        <translation>0 0.2 0</translation>
+        <extents>0.2 0.01 0.01</extents>
+      </geom>
+      <geom type="box">
+        <translation>0 -0.2 0</translation>
+        <extents>0.2 0.01 0.01</extents>
+      </geom>
+    </body>
+  </KinBody>
+</Environment>
+'''
+        env=self.env
+        self.LoadDataEnv(xmldata)
+        env.Save('test_writekinematicsonly.dae',Environment.SelectionOptions.Everything,{'externalref':'*'})
+        env2 = Environment()
+        env2.Load('test_writekinematicsonly.dae')
+        assert(len(env.GetBodies())==len(env2.GetBodies()))
+        for body in env.GetBodies():
+            self.log.info(body.GetName())
+            body2 = env2.GetKinBody(body.GetName())
+            misc.CompareBodies(body,body2,epsilon=g_epsilon)
+
     def test_writekinematicsonly(self):
         self.log.info('test writing kinematics only')
         env=self.env